scripts/spectrum/requirements.txt
pdfplumber>=0.10.0
playwright>=1.40.0
python-dotenv>=1.0.0
requests>=2.31.0
imapclient>=2.3.0
